consumers/requirements.txt
# consumers/requirements.txt
kafka-python
pymongo
